+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Config subtabs
 - CharacterFlag Multiplier Selector
 - Presets
+- Separators for Capacity UI
 
 ## [1.7.0]